+++ lib/types/form-types.ts
@@ -133,7 +133,7 @@ export type FormRecord = {
   form: FormProperties;
   securityAttribute: string;
   reCaptchaID?: string;
-  updatedAt: string;
+  updatedAt?: string;
   [key: string]: string | boolean | SubmissionProperties | FormProperties | undefined;
 };
